backend/requirements.txt
Flask==3.0.0
Flask-CORS==4.0.0
google-auth==2.25.0
google-auth-oauthlib==1.2.0
PyJWT==2.8.0
python-dotenv==1.0.0
